recontoolkit-model/src/main/java/eu/delpeuch/antonin/recontoolkit/protocol/SuggestResponse.java

package eu.delpeuch.antonin.recontoolkit.protocol;

/*-
 * #%L
 * ReconToolkit data model
 * %%
 * Copyright (C) 2022 - 2023 ReconToolkit Developers
 * %%
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *      http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * #L%
 */

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonProperty;

import java.util.Collections;
import java.util.List;
import java.util.Objects;

/**
 * A list of entities, properties or types meant to be displayed as a list of suggestions following input from the user
 * in a text field (for autocompletion and validation purposes).
 */
public class SuggestResponse<T> {

    private final List<T> results;

    @JsonCreator
    public SuggestResponse(
            @JsonProperty("result") List<T> results) {
        this.results = results == null ? Collections.emptyList() : results;
    }

    @JsonProperty("result")
    public List<T> getResults() {
        return results;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        SuggestResponse<?> that = (SuggestResponse<?>) o;
        return results.equals(that.results);
    }

    @Override
    public int hashCode() {
        return Objects.hash(results);
    }

    @Override
    public String toString() {
        return "SuggestResponse{" +
                "results=" + results +
                '}';
    }
}
